states: Handle state change callbacks in State trait object
There was a design issue of whether callbacks should be handled on the
`states::run` function or in a method of the `states::State` trait
(that I call `handle_callbacks`). The latter seems the right choice,
in order to not clutter `run`.

I decided to implement it only for `State` trait objects, because each
state hopefully doesn't need to override how callbacks are
handled. This avoids bloating the binary with one copy for each state.

`State::handle` returns a `StateTransitioner` instead of a state: the
transitioner can either perform the transition (returning the next
state) or cancel it (returning the previous state). This was needed to
work around the borrow checker.

PS: this is just to show how the code currently is; the code doesn't
compile yet due to a borrow checker error (to solve it
`StateTransitioner` will need to stop holding a closure). Also,
there's a test commented out on src/states/mod.rs